+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.1.2
+* Fix 'select' method in MockISelectionManager to allow deselecting selection
+* Check if 'callback' is defined in MockISelectionManager
+* Add tests for MockISelectionManager
 
 ## 6.1.1
 * powerbi-visuals-api update to 5.9.0